website/requirements.txt
plotly
mysql-connector-python
streamlit